Rename juno.command to juno.commands
